Renamed packages and change path to games/popular
