+++ backend/src/main/java/com/zpi/backend/exception_handlers/ExceptionController.java
@@ -1,14 +1,14 @@
 package com.zpi.backend.exception_handlers;
 
-import com.zpi.backend.gameInstance.GameInstanceDoesNotExistException;
-import com.zpi.backend.gameInstance.GameInstanceStatusException;
+import com.zpi.backend.game_instance.GameInstanceDoesNotExistException;
+import com.zpi.backend.game_instance.GameInstanceStatusException;
 import com.zpi.backend.category.CategoryAlreadyExistsException;
 import com.zpi.backend.category.CategoryDoesNotExistException;
 import com.zpi.backend.game.GameAlreadyAcceptedException;
 import com.zpi.backend.game.GameAlreadyExistsException;
 import com.zpi.backend.game.GameAlreadyRejectedException;
 import com.zpi.backend.game.GameDoesNotExistException;
-import com.zpi.backend.gameInstanceImage.GameInstanceImageDoesNotExistException;
+import com.zpi.backend.game_instance_image.GameInstanceImageDoesNotExistException;
 import com.zpi.backend.user.UndefinedUserException;
 import com.zpi.backend.user.UserAlreadyExistsException;
 import com.zpi.backend.user.UserDoesNotExistException;
