+++ backend/src/main/java/com/zpi/backend/game_instance/GameInstanceDoesNotExistException.java
@@ -1,4 +1,4 @@
-package com.zpi.backend.gameInstance;
+package com.zpi.backend.game_instance;
 
 public class GameInstanceDoesNotExistException extends Exception{
     public GameInstanceDoesNotExistException(String message) {
