+++ backend/src/main/java/com/zpi/backend/game_instance/GameInstanceListDTO.java
@@ -1,9 +1,6 @@
-package com.zpi.backend.gameInstance;
+package com.zpi.backend.game_instance;
 
-import com.zpi.backend.game.GameDTO;
 import com.zpi.backend.game.GameListDTO;
-import com.zpi.backend.user.UserGameDTO;
-import com.zpi.backend.user.UserGameGuestDTO;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
