+++ backend/src/main/java/com/zpi/backend/game_instance/GameInstanceSpecification.java
@@ -1,4 +1,4 @@
-package com.zpi.backend.gameInstance;
+package com.zpi.backend.game_instance;
 
 import com.zpi.backend.category.Category;
 import com.zpi.backend.game.Game;
