+++ backend/src/main/java/com/zpi/backend/game_instance/GameInstanceStatusException.java
@@ -1,4 +1,4 @@
-package com.zpi.backend.gameInstance;
+package com.zpi.backend.game_instance;
 
 public class GameInstanceStatusException extends Exception{
     public GameInstanceStatusException(String message) {
