+++ backend/src/main/java/com/zpi/backend/game_instance/NewGameInstanceDTO.java
@@ -1,4 +1,4 @@
-package com.zpi.backend.gameInstance;
+package com.zpi.backend.game_instance;
 
 import lombok.Data;
 import lombok.EqualsAndHashCode;
