+++ backend/src/main/java/com/zpi/backend/game_instance_image/GameInstanceImageDoesNotExistException.java
@@ -1,4 +1,4 @@
-package com.zpi.backend.gameInstanceImage;
+package com.zpi.backend.game_instance_image;
 
 public class GameInstanceImageDoesNotExistException extends Exception{
     public GameInstanceImageDoesNotExistException(String message) {
